refactor: extract canonical chat timeline
